fix(release): resolve platform sidecars on consumer install
