+++ scripts/prepare-senpi-bundled-workspaces.mjs
@@ -1,7 +1,16 @@
 #!/usr/bin/env node
-import { cpSync, existsSync, mkdirSync, readdirSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import { cpSync, existsSync, mkdirSync, readFileSync, rmSync } from "node:fs";
 import { dirname, join, relative, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
+import { stagePublishManifest } from "./prepare-senpi-publish-manifest.mjs";
+export {
+	bundlablePublishPackageNames,
+	isPlatformConstrainedPackage,
+	listStagedPublishPackageNames,
+	ownedRegistryAliases,
+	rewriteOwnedRegistryAliases,
+	stagePublishManifest,
+} from "./prepare-senpi-publish-manifest.mjs";
 
 const root = fileURLToPath(new URL("..", import.meta.url));
 
@@ -47,29 +56,6 @@ const bundledWorkspaces = [
 	},
 ];
 const internalPackageNames = new Set(bundledWorkspaces.map((workspace) => workspace.packageName));
-export const ownedRegistryAliases = new Map([
-	["@earendil-works/pi-ai", "@code-yeongyu/senpi-ai"],
-	["@earendil-works/pi-agent-core", "@code-yeongyu/senpi-agent-core"],
-	["@earendil-works/pi-tui", "@code-yeongyu/senpi-tui"],
-	["@earendil-works/pi-pty", "@code-yeongyu/senpi-pty"],
-]);
-
-export function rewriteOwnedRegistryAliases(manifest) {
-	for (const dependencyField of ["dependencies", "optionalDependencies"]) {
-		const dependencies = manifest[dependencyField];
-		if (!dependencies) {
-			continue;
-		}
-		for (const [packageName, aliasName] of ownedRegistryAliases) {
-			const version = dependencies[packageName];
-			if (typeof version === "string" && !version.startsWith("npm:")) {
-				dependencies[packageName] = `npm:${aliasName}@${version}`;
-			}
-		}
-	}
-	return manifest;
-}
-
 function requiredFilesForWorkspace(workspace, nativeTargets) {
 	const requiredFiles = [...(workspace.requiredFiles ?? ["package.json", "dist/index.js"])];
 	if (workspace.nativePrebuild) {
@@ -111,109 +97,6 @@ export function directNodeModulesPackageName(lockPath) {
 		return parts.length === 2 ? `${parts[0]}/${parts[1]}` : undefined;
 	}
 	return parts.length === 1 ? parts[0] : undefined;
-}
-
-export function listStagedPublishPackageNames(codingAgentNodeModules) {
-	const packageNames = [];
-	for (const entry of readdirSync(codingAgentNodeModules, { withFileTypes: true })) {
-		if (!entry.isDirectory() || entry.name.startsWith(".")) {
-			continue;
-		}
-		if (entry.name.startsWith("@")) {
-			const scopeDir = join(codingAgentNodeModules, entry.name);
-			for (const scoped of readdirSync(scopeDir, { withFileTypes: true })) {
-				if (scoped.isDirectory()) {
-					packageNames.push(`${entry.name}/${scoped.name}`);
-				}
-			}
-			continue;
-		}
-		packageNames.push(entry.name);
-	}
-	return packageNames.sort((a, b) => a.localeCompare(b));
-}
-
-// npm evaluates `os`/`cpu`/`libc` against the INSTALLING machine, but bundleDependencies ships a
-// single tarball to every platform and npm republishes the bundled set as required `dependencies`
-// in the registry manifest. Bundling a native binary that only matches the publish runner
-// (linux-x64 in publish-npm.yml) therefore aborts `npm install @code-yeongyu/senpi` with
-// EBADPLATFORM on every other platform, before a single file is written. Such packages stay
-// ordinary optional registry deps: npm resolves the binary matching the installing platform, and
-// a failed fetch degrades the feature instead of failing the install.
-export function isPlatformConstrainedPackage(packageDir) {
-	let manifest;
-	try {
-		manifest = JSON.parse(readFileSync(join(packageDir, "package.json"), "utf8"));
-	} catch {
-		// An unreadable staged package cannot be judged here; leave it bundled so the existing
-		// staging and pack gates report it instead of this filter dropping it silently.
-		return false;
-	}
-	return ["os", "cpu", "libc"].some((field) => {
-		// npm accepts a bare string as well as the documented array form.
-		const constraint = manifest[field];
-		if (typeof constraint === "string") {
-			return constraint.length > 0;
-		}
-		return Array.isArray(constraint) && constraint.length > 0;
-	});
-}
-
-export function bundlablePublishPackageNames(codingAgentNodeModules, packageNames) {
-	return packageNames.filter((name) => !isPlatformConstrainedPackage(join(codingAgentNodeModules, name)));
-}
-
-// The publish tarball must be fully self-contained: every runtime dependency edge in
-// the coding-agent manifest (registry deps AND the 5 vendored workspace packages) is
-// staged into packages/coding-agent/node_modules, and bundleDependencies lists every
-// portable staged package (platform-constrained natives are excluded — see
-// isPlatformConstrainedPackage). npm then needs no registry fetch for anything that
-// installs identically on every platform at install time. The historical
-// partial bundle (only the 5 workspace packages + their closure) forced npm to fetch
-// the other runtime deps from the registry, where arborist nondeterministically tried
-// to resolve the registry-absent `^2026.x` workspace specs (ETARGET) and aborted reify
-// mid-flight, leaving arbitrary deps (cross-spawn, which, @modelcontextprotocol/sdk)
-// missing from the installed CLI (ERR_MODULE_NOT_FOUND).
-export function stagePublishManifest(repoRoot) {
-	const codingAgentDir = join(repoRoot, "packages/coding-agent");
-	const manifestPath = join(codingAgentDir, "package.json");
-	const manifest = JSON.parse(readFileSync(manifestPath, "utf8"));
-	const codingAgentNodeModules = join(codingAgentDir, "node_modules");
-	const stagedPackageNames = listStagedPublishPackageNames(codingAgentNodeModules);
-	const stagedSet = new Set(stagedPackageNames);
-
-	const runtimeDependencyFields = ["dependencies", "optionalDependencies"];
-	const missing = [];
-	for (const field of runtimeDependencyFields) {
-		for (const [name, spec] of Object.entries(manifest[field] ?? {})) {
-			if (/^(file|link|workspace):/.test(spec)) {
-				throw new Error(
-					`packages/coding-agent/package.json ${field}.${name} uses a local spec (${spec}); the published tarball must not reference local paths.`,
-				);
-			}
-			if (!stagedSet.has(name)) {
-				missing.push(name);
-			}
-		}
-	}
-	if (missing.length > 0) {
-		throw new Error(
-			`packages/coding-agent/node_modules is missing staged runtime dependencies: ${missing.join(", ")}. Run npm install before publishing.`,
-		);
-	}
-
-	const bundlablePackageNames = bundlablePublishPackageNames(codingAgentNodeModules, stagedPackageNames);
-	// Keep the original dependency keys so npm packs the modules at the import paths
-	// the compiled source uses. Resolve those keys through fork-owned aliases instead
-	// of attempting to fetch lockstep versions from the upstream-owned namespace.
-	rewriteOwnedRegistryAliases(manifest);
-	manifest.bundleDependencies = bundlablePackageNames;
-	// npm accepts both spellings; the checked-in manifest carries both, so keep them in sync.
-	if (manifest.bundledDependencies !== undefined) {
-		manifest.bundledDependencies = [...bundlablePackageNames];
-	}
-	writeFileSync(manifestPath, `${JSON.stringify(manifest, null, "\t")}\n`);
-	return bundlablePackageNames;
 }
 
 export function copyPublishDependencies(repoRoot) {
